+++ src/forceCode.d.ts
@@ -151,19 +151,13 @@ export interface IForceService {
   storageRoot: string;
   describe?: IMetadataDescribe;
   containerId?: string;
-  statusTimeout: any;
   containerMembers: IContainerMember[];
   containerAsyncRequestId?: string;
   conn: Connection;
-  outputChannel: vscode.OutputChannel;
-  statusBarItem: vscode.StatusBarItem;
   fcDiagnosticCollection: vscode.DiagnosticCollection;
   uuid: string;
   connect(): Promise<IForceService>;
   newContainer(force: Boolean): Promise<IForceService>;
-  showStatus(message: string): void;
-  clearLog(): void;
-  resetStatus(): void;
   checkForFileChanges(): any;
 }